0.0.6 PATCH
Introduced fixed filling, i.e. canonical ensemble, method.
This activated using ed_filling/=0 (integer) to select which sectors
to address (reducing the computational cost).

Separated the chemical potential and Hartree shifts contributions and
deleted the additive constants which affected calculations at large
values of Uloc.

Added full Kondo terms evaluation in the local energy.
Checked sign of the coupling (-Jk s.S):
Jk>0 --> FM
Jk<0 --> AFM

Testing.

>> NOTE:
The Kondo models, i.e. systems with Nimp spin-1/2-like impurities,
is improperly coded as a fermionic state. Although this is not necessarily
wrong and helps interpolating between Anderson model and Kondo one, it
requires states which do not need to contribute, empty or doubly occupied.
These can be projected out using a large value of U. However the groundstate
properties should be (in fact they are) insensitive to this choice.

A better description should feature a more elaborated constraints, e.g.
impose that Nimp_up + Nimp_dw = Nimp. This however breaks down the
separability of the problem in Up and Dw electrons, which is at the hearth
of this implementation.
This would rather require use of a single map H from the sectors S(Nup,Ndw) to Fock
space, rather than two separate maps Hup, Hdw, which is possible only through a
serious rewriting of the code. Possible, but hard and annoying.

Should this problem become more serious we can think of implementing this
part at the cost of a decreased performances in the ED solution.