Intermediate commit.
Starting coding the Kondo algorithm using hard spin, as opposed
to fermion.
Stop because I realized that a possible better way to deal with
this problem is to use fermions and project out unwanted states in
the sectors, i.e. build sectors which have encoded the condition
that fully empty and fully occupied states are un-accessible.
yet such states should exist in the Fock space.