2.2.2 BUG FIX
Remove a bug in all H*_kondo which prevented correct evaluation of the
matrix element. Htmp  was not initialized to zero yet an implicit
summation, inherited from the old version was performed.
This led to bad evaluation of the matrix element, bad eigenenergies
and then to bad density distribution among the site.

Further testing.