Start development of lattice density matrix [`add_features` branch]
Here we rebase `add_features` branch onto current master, so to add an
(optional) evaluation of the full density matrix for the lattice. Both
for the ground(lanc) and thermal states (arpack or full ed_method).

No support for Kondo impurities is planned for now (we can discuss, but
I don't need to deal with them for now... I'd either add suitable input
checks or leave it there for someone else to test the Nimp > 0 case).

This /draft/ version compiles and runs with no segfaults (as tested on
the `2d_plaquette_ed` driver in the database, plus a custom driver for
the Hubbard dimer that I'm preparing). No I/O procedures have been set
up for the density matrix so no check on the results has been done yet.

------------------------------------------------------------------------

Side notes on the current functionality on master:

1. I wonder why offdiag_gf_flag = offdiag_gf_flag .AND. Norb>1, at line
   187 of SETUP.f90. For now I commented away the statement, for that I
   need to inspect cross-site components of the dimer/plaquette GFs. If
   there is a reason for the constraint just inform me and we will find
   a clean way to accommodate the different needs in the input file.

2. Hartree energies are computed with respect to a different zero, than
   what we do in the impurity solvers (EDIpack, LIB_DMFT_ED, CDMFT_ED).
   Namely the terms proportional to Uloc,Ust,(Ust-Jh) only (no <nᵢnⱼ>)
   here are discarded. This means that hfmode=true amounts to different
   conventions in different codes, making for difficult comparisons.
   Am I missing something? I see vestigia of those contributions in the
   source code (as comments) here and there, so I think they were pruned
   intentionally. For now I'm leaving this untouched, to keep the branch
   focused and avoid unintended regressions.

3. I have not investigated the issue at all, but there seem to be race
   conditions in I/O operations: running under mpi leads /occasionally/*
   to write errors (drude.ed file cannot be opened cause nonexistent).
   As said, I did not spend time for now on this, so don't know whether
   the problem is limited to the OC evaluation per se' or it's broader.

   */ With "occasionally" I mean that doing a full swipe on Uloc values,
      only some calculation did stop with a I/O error. All other points
      did run fine, with very reasonable results. Running serially the
      issue disappears completely. I may give it a look, but no promise
      on the time horizon: I'll proceed on the density matrix work with
      serial runs only, for now.